Upgraded AGP to 9.3.0 and pinned Kotlin and the Compose plugin to 2.3.21
Upgraded the Compose BOM to 2026.06.01

Migrated off the deprecated MenuAnchorType and LocalClipboardManager APIs

Removed redundant Float conversions and an unnecessary safe call

Added mirakle.gradle to .gitignore

+++ app/src/main/java/org/cygnusx1/openbu/ui/RtspStreamScreen.kt
@@ -54,8 +54,8 @@ fun RtspStreamScreen(player: ExoPlayer?, isRelayed: Boolean = false) {
                     } else {
                         val maxX = (size.width * (newScale - 1)) / 2
                         val maxY = (size.height * (newScale - 1)) / 2
-                        offsetX.floatValue = (offsetX.floatValue + pan.x).coerceIn(-maxX.toFloat(), maxX.toFloat())
-                        offsetY.floatValue = (offsetY.floatValue + pan.y).coerceIn(-maxY.toFloat(), maxY.toFloat())
+                        offsetX.floatValue = (offsetX.floatValue + pan.x).coerceIn(-maxX, maxX)
+                        offsetY.floatValue = (offsetY.floatValue + pan.y).coerceIn(-maxY, maxY)
                     }
                 }
             },
